build: Generate a JSON SPDX v2.3 SBOM for Qt Creator and plugins
The Python building scripts will now pass two additional options to
the build system:
- to enable JSON SPDX SBOM generation when the general SBOM option is
  enabled
- to error out if the JSON generation fails

Previously JSON generation was not enabled due to an issue with
handling spaces in Python interpreter paths on Windows.
The issue has been fixed in newer Qt versions (6.8.5, 6.9.3, 6.10.1,
6.11.0).

The CMake build system was modified to detect a Qt version that does
not support JSON SPDX generation, in which case it will forcefully
disable JSON generation and show a message about it.
This allows switching to an older Qt version in CI (if needed) without
breaking the build.

The external subprojects were modified to pass through the relevant
options as well.

SPDX JSON generation requires the SBOM Python interpreter to have the
pyspdxtools module installed, which is installed in Qt's provisioning
scripts, and is already available in Qt's CI.

The build_plugin.py script was modified to support the new options,
as well as add the no_sbom option to be in sync with the build.py
script.

Depends on 60fd81ce645ca2bfe402b4439983bac228859b29 in qtbase.

Change-Id: I284fdb996ff4c749e9a5da124c72501ba94e0744